requirements.txt
ultralytics>=8.0.0
manga-ocr>=0.1.0
simple-lama-inpainting>=0.1.0
Pillow>=9.0.0
requests>=2.28.0
opencv-python>=4.7.0
numpy<2.0.0
scikit-learn>=1.3.0
torch>=2.6.0
